Setting up a multidir project: makefile improvements
The following commit adjusts project to new, multidirectory file structure.
The main improvements are based on using variables to hold each directory
separetely and wildcard and patsubst macros to include all files of a given
extension in single line without any changes as long as those are contained
in the same directory.